Fixed a fatal parsing bug where certain (valid) syntax was not parsed correctly and it wasn't possible to correctly call functions. Functions may now return and take tuples as arguments. Introduced the concept of a dynamic type environment for dynamically created types (tuples, arrays, generics, etc.)

+++ AAVM.cpp
@@ -646,7 +646,7 @@ void AAVM::exec(AAProgram::Procedure* procedure, aa::stack<AARuntimeEnvironment>
 			int count = AAVM_GetArgument(0);
 			aa::array<AAPrimitiveType> types = aa::array<AAPrimitiveType>(count);
 			aa::array<AAVal> values = aa::array<AAVal>(count);
-			for (size_t i = 0; i < count; i++) {
+			for (int i = 0; i < count; i++) {
 				types[count - i - 1] = (AAPrimitiveType)AAVM_GetArgument(i + 1);
 				values[count - i - 1] = aa::vm::PopSomething(types[count - i - 1], stack).as_val();
 			}
